requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

Flask==2.2.2
numpy==1.24.2
Pillow==9.0.1
python-dotenv==0.21.1
requests==2.27.1
torch==1.13.1
torchvision==0.14.1
waitress==2.1.2
Werkzeug==2.2.2
gunicorn==20.1.0
ultralytics==8.0.84

